Renamed FieldRenderer component
