+++ packages/roqueform/src/main/index.ts
@@ -1,7 +1,8 @@
 export * from './AccessorContext';
 export * from './applyPlugins';
 export * from './createField';
-export * from './Field';
+export * from './field-types';
+export * from './FieldRenderer';
 export * from './objectAccessor';
 export * from './useField';
 export * from './validationPlugin';
