+++ packages/roqueform/src/main/useField.ts
@@ -2,12 +2,12 @@ import { useContext, useRef } from 'react';
 import { AccessorContext } from './AccessorContext';
 import { createField } from './createField';
 import { callOrGet } from './utils';
-import { Field, Plugin } from './Field';
+import { Field, Plugin } from './field-types';
 
 /**
  * Creates the new field.
  *
- * @returns The `Field` instance.
+ * @returns The {@linkcode Field} instance.
  * @template T The value controlled by the field.
  */
 export function useField<T = any>(): Field<T | undefined>;
@@ -16,7 +16,7 @@ export function useField<T = any>(): Field<T | undefined>;
  * Creates the new field.
  *
  * @param initialValue The initial value assigned to the field.
- * @returns The `Field` instance.
+ * @returns The {@linkcode Field} instance.
  * @template T The value controlled by the field.
  */
 export function useField<T>(initialValue: T | (() => T)): Field<T>;
@@ -26,7 +26,7 @@ export function useField<T>(initialValue: T | (() => T)): Field<T>;
  *
  * @param initialValue The initial value assigned to the field.
  * @param plugin Enhances the field with additional functionality.
- * @returns The `Field` instance.
+ * @returns The {@linkcode Field} instance.
  * @template T The value controlled by the field.
  * @template P The enhancement added by the plugin.
  */
