+++ packages/roqueform/src/test/FieldRenderer.test.tsx
@@ -1,23 +1,23 @@
 import React from 'react';
 import { act, render } from '@testing-library/react';
-import { Field, useField } from '../main';
+import { Field, FieldRenderer, useField } from '../main';
 
-describe('Field', () => {
+describe('FieldRenderer', () => {
   test('passes the field as an argument', () => {
     const Test = () => {
       const rootField = useField({ foo: 'bar' });
 
       return (
-        <Field field={rootField.at('foo')}>
+        <FieldRenderer field={rootField.at('foo')}>
           {field => {
             expect(field).toBe(rootField.at('foo'));
             return null;
           }}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
   });
 
   test('re-renders if field value is changed externally', async () => {
@@ -28,16 +28,16 @@ describe('Field', () => {
       rootField = useField();
 
       return (
-        <Field field={rootField}>
+        <FieldRenderer field={rootField}>
           {() => {
             renderCount++;
             return null;
           }}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
 
     await act(() => rootField.setValue(111));
 
@@ -52,16 +52,16 @@ describe('Field', () => {
       rootField = useField();
 
       return (
-        <Field field={rootField}>
+        <FieldRenderer field={rootField}>
           {() => {
             renderCount++;
             return null;
           }}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
 
     await act(() => rootField.notify());
 
@@ -76,16 +76,16 @@ describe('Field', () => {
       rootField = useField({ foo: 111 });
 
       return (
-        <Field field={rootField}>
+        <FieldRenderer field={rootField}>
           {() => {
             renderCount++;
             return null;
           }}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
 
     await act(() => rootField.at('foo').setValue(222));
 
@@ -100,7 +100,7 @@ describe('Field', () => {
       rootField = useField({ foo: 111 });
 
       return (
-        <Field
+        <FieldRenderer
           field={rootField}
           eagerlyUpdated={true}
         >
@@ -108,11 +108,11 @@ describe('Field', () => {
             renderCount++;
             return null;
           }}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
 
     await act(() => rootField.at('foo').setValue(222));
 
@@ -127,16 +127,16 @@ describe('Field', () => {
       rootField = useField({ foo: 111 });
 
       return (
-        <Field
+        <FieldRenderer
           field={rootField.at('foo')}
           onChange={handleChangeMock}
         >
           {() => null}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
 
     await act(() => rootField.at('foo').setValue(222));
 
@@ -152,16 +152,16 @@ describe('Field', () => {
       rootField = useField({ foo: 111 });
 
       return (
-        <Field
+        <FieldRenderer
           field={rootField.at('foo')}
           onChange={handleChangeMock}
         >
           {() => null}
-        </Field>
+        </FieldRenderer>
       );
     };
 
-    render(<Test />);
+    render(<Test/>);
 
     await act(() => {
       rootField.at('foo').setTransientValue(222);
